split http with tcp code
